refactor: move application path helpers to domain
