+++ src/anteumbra/application/path_service.py
@@ -1,5 +1,5 @@
 """Stable application API for runtime path normalization."""
 
-from anteumbra.infrastructure.utils.path_utils import normalize_path, path_to_key
+from anteumbra.domain.paths import normalize_path, path_to_key
 
 __all__ = ["normalize_path", "path_to_key"]
